Added in test case and code mod
