+++ src/mutations/typeMutating/createNonNullAssertion.ts
@@ -2,7 +2,7 @@ import { ITextInsertMutation, ITextSwapMutation } from "automutate";
 import * as ts from "typescript";
 
 import { FileMutationsRequest } from "../../mutators/fileMutator";
-import { findAliasOfType } from "../aliasing/findAliasOfType";
+import { findAliasOfType } from "../aliasing/findAliasOfTypes";
 
 // The following node types need to be wrapped in parenthesis to stop the ! from being applied to the wrong (last) element:
 // As expressions: foo as Bar
