+++ src/mutators/collectMutationsFromNodes.ts
@@ -40,7 +40,6 @@ const tryGetMutation = <TNode extends ts.Node>(request: FileMutationsRequest, no
     try {
         return visitor(node, request);
     } catch (error) {
-        debugger;
         request.options.logger.stderr.write(
             `\nError in ${request.sourceFile.fileName} at node '${node.getText(request.sourceFile)}' (position ${node.pos}):\n\t`,
         );
