+++ test/cases/unit/implicit any/aliasing/expected.ts
@@ -1,0 +1,7 @@
+(function () {
+    const receivesString = (text: string /* todo */) => {
+        return text;
+    }
+
+    receivesString("");
+})();
