+++ test/cases/unit/implicit any/aliasing/tsconfig.json
@@ -1,0 +1,6 @@
+{
+    "compilerOptions": {
+        "strictNullChecks": true
+    },
+    "files": ["actual.ts"]
+}
