+++ test/cases/unit/implicit any/original.ts
@@ -1,0 +1,7 @@
+(function () {
+    const receivesString = (text) => {
+        return text;
+    }
+
+    receivesString("");
+})();
